Extract Bonsplit tab bar debug support
