+++ Sources/TabManager.swift
@@ -5350,13 +5350,6 @@ class TabManager: ObservableObject {
         }
     }
 
-    /// Refresh Bonsplit right-side action button backdrop tuning for all workspaces.
-    func refreshSplitButtonBackdropEffect() {
-        for workspace in tabs {
-            workspace.refreshSplitButtonBackdropEffect()
-        }
-    }
-
     func applySurfaceTabBarButtons(
         _ buttons: [CmuxSurfaceTabBarButton],
         sourcePath: String?,
