+++ Sources/Workspace.swift
@@ -74,119 +74,6 @@ private final class WorkspacePendingTerminalInputObserver: @unchecked Sendable {
     var observer: NSObjectProtocol?
 }
 
-struct BonsplitTabBarDebugNumberSetting {
-    let key: String
-    let defaultValue: Double
-    let range: ClosedRange<Double>
-    let step: Double
-
-    func resolved(_ value: Double) -> Double {
-        guard value.isFinite else { return defaultValue }
-        return min(max(value, range.lowerBound), range.upperBound)
-    }
-
-    func currentValue(defaults: UserDefaults = .standard) -> Double {
-#if DEBUG
-        guard defaults.object(forKey: key) != nil else {
-            return defaultValue
-        }
-        return resolved(defaults.double(forKey: key))
-#else
-        return defaultValue
-#endif
-    }
-
-    func format(_ value: Double) -> String {
-        String(format: "%.3f", resolved(value))
-    }
-}
-
-enum BonsplitTabBarDebugSettings {
-    static let backdropFadeWidth = 99.75
-
-    static let separatorFadeWidthSetting = BonsplitTabBarDebugNumberSetting(
-        key: "debugBonsplitTabBarSeparatorFadeWidthV2",
-        defaultValue: backdropFadeWidth,
-        range: 0.0...140.0,
-        step: 0.25
-    )
-    static let contentFadeWidthSetting = BonsplitTabBarDebugNumberSetting(
-        key: "debugBonsplitTabBarContentFadeWidth",
-        defaultValue: 28.875,
-        range: 0.0...80.0,
-        step: 0.5
-    )
-    static let solidSurfaceWidthAdjustmentSetting = BonsplitTabBarDebugNumberSetting(
-        key: "debugBonsplitTabBarSolidSurfaceWidthAdjustmentV2",
-        defaultValue: -80.0,
-        range: -80.0...120.0,
-        step: 0.5
-    )
-
-    static let separatorFadeWidthKey = separatorFadeWidthSetting.key
-    static let contentFadeWidthKey = contentFadeWidthSetting.key
-    static let solidSurfaceWidthAdjustmentKey = solidSurfaceWidthAdjustmentSetting.key
-    static let defaultSeparatorFadeWidth = separatorFadeWidthSetting.defaultValue
-    static let defaultContentFadeWidth = contentFadeWidthSetting.defaultValue
-    static let defaultSolidSurfaceWidthAdjustment = solidSurfaceWidthAdjustmentSetting.defaultValue
-    static let separatorFadeWidthRange = separatorFadeWidthSetting.range
-    static let contentFadeWidthRange = contentFadeWidthSetting.range
-    static let solidSurfaceWidthAdjustmentRange = solidSurfaceWidthAdjustmentSetting.range
-    static let separatorFadeWidthStep = separatorFadeWidthSetting.step
-    static let contentFadeWidthStep = contentFadeWidthSetting.step
-    static let solidSurfaceWidthAdjustmentStep = solidSurfaceWidthAdjustmentSetting.step
-
-    static func resolvedSeparatorFadeWidth(_ width: Double) -> Double {
-        separatorFadeWidthSetting.resolved(width)
-    }
-
-    static func resolvedContentFadeWidth(_ width: Double) -> Double {
-        contentFadeWidthSetting.resolved(width)
-    }
-
-    static func resolvedSolidSurfaceWidthAdjustment(_ width: Double) -> Double {
-        solidSurfaceWidthAdjustmentSetting.resolved(width)
-    }
-
-    static func separatorFadeWidth(defaults: UserDefaults = .standard) -> Double {
-        separatorFadeWidthSetting.currentValue(defaults: defaults)
-    }
-
-    static func contentFadeWidth(defaults: UserDefaults = .standard) -> Double {
-        contentFadeWidthSetting.currentValue(defaults: defaults)
-    }
-
-    static func solidSurfaceWidthAdjustment(defaults: UserDefaults = .standard) -> Double {
-        solidSurfaceWidthAdjustmentSetting.currentValue(defaults: defaults)
-    }
-
-    static func formatPixels(_ value: Double) -> String {
-        String(format: "%.3f", value)
-    }
-
-    static func currentTuningDescription(defaults: UserDefaults = .standard) -> String {
-        let effect = Workspace.bonsplitSplitButtonBackdropEffect(defaults: defaults)
-        return [
-            "bonsplit-tabbar-tuning",
-            "separatorFadeWidth=\(formatPixels(separatorFadeWidth(defaults: defaults)))",
-            "contentFadeWidth=\(formatPixels(contentFadeWidth(defaults: defaults)))",
-            "solidSurfaceWidthAdjustment=\(formatPixels(solidSurfaceWidthAdjustment(defaults: defaults)))",
-            "fadeWidth=\(String(format: "%.3f", Double(effect.fadeWidth)))",
-            "solidWidth=\(String(format: "%.3f", Double(effect.solidWidth)))",
-            "fadeRampStartFraction=\(String(format: "%.3f", Double(effect.fadeRampStartFraction)))",
-            "trailingOpacity=\(String(format: "%.3f", Double(effect.trailingOpacity)))",
-            "contentOcclusionFraction=\(String(format: "%.3f", Double(effect.contentOcclusionFraction)))",
-            "masksTabContent=\(effect.masksTabContent ? "true" : "false")"
-        ].joined(separator: " ")
-    }
-
-    static func copyCurrentTuningToPasteboard(defaults: UserDefaults = .standard) {
-        let pasteboard = NSPasteboard.general
-        pasteboard.clearContents()
-        pasteboard.setString(currentTuningDescription(defaults: defaults), forType: .string)
-    }
-}
-
 func cmuxSurfaceContextName(_ context: ghostty_surface_context_e) -> String {
     switch context {
     case GHOSTTY_SURFACE_CONTEXT_WINDOW:
@@ -7693,28 +7580,6 @@ final class Workspace: Identifiable, ObservableObject {
             splitButtonBackdropHex: surfaceHex,
             paneBackgroundHex: paneBackgroundHex,
             borderHex: borderHex
-        )
-    }
-
-    nonisolated static let bonsplitSplitButtonBackdropSoftness: CGFloat = 0.60
-
-    nonisolated static func bonsplitSplitButtonBackdropEffect(
-        defaults: UserDefaults = .standard
-    ) -> BonsplitConfiguration.Appearance.SplitButtonBackdropEffect {
-        .init(
-            style: .translucentChrome,
-            fadeWidth: CGFloat(BonsplitTabBarDebugSettings.backdropFadeWidth),
-            contentFadeWidth: CGFloat(BonsplitTabBarDebugSettings.contentFadeWidth(defaults: defaults)),
-            solidWidth: 23.875,
-            solidSurfaceWidthAdjustment: CGFloat(
-                BonsplitTabBarDebugSettings.solidSurfaceWidthAdjustment(defaults: defaults)
-            ),
-            separatorFadeWidth: CGFloat(BonsplitTabBarDebugSettings.separatorFadeWidth(defaults: defaults)),
-            fadeRampStartFraction: bonsplitSplitButtonBackdropSoftness,
-            leadingOpacity: 0,
-            trailingOpacity: 0.8625,
-            contentOcclusionFraction: 0.6875,
-            masksTabContent: true
         )
     }
