+++ Sources/cmuxApp.swift
@@ -3350,206 +3350,6 @@ private struct SplitButtonLayoutDebugView: View {
     }
 }
 
-// MARK: - Bonsplit Tab Bar Debug Window
-
-private final class BonsplitTabBarDebugWindowController: NSWindowController, NSWindowDelegate {
-    static let shared = BonsplitTabBarDebugWindowController()
-
-    private init() {
-        let window = NSPanel(
-            contentRect: NSRect(x: 0, y: 0, width: 520, height: 420),
-            styleMask: [.titled, .closable, .utilityWindow],
-            backing: .buffered,
-            defer: false
-        )
-        window.title = String(
-            localized: "debug.bonsplitTabBarDebug.title",
-            defaultValue: "Bonsplit Tab Bar Debug"
-        )
-        window.titleVisibility = .visible
-        window.titlebarAppearsTransparent = false
-        window.isMovableByWindowBackground = true
-        window.isReleasedWhenClosed = false
-        window.identifier = NSUserInterfaceItemIdentifier("cmux.bonsplitTabBarDebug")
-        window.center()
-        window.contentView = NSHostingView(rootView: BonsplitTabBarDebugView())
-        AppDelegate.shared?.applyWindowDecorations(to: window)
-        super.init(window: window)
-        window.delegate = self
-    }
-
-    @available(*, unavailable)
-    required init?(coder: NSCoder) { fatalError() }
-
-    func show() {
-        window?.center()
-        window?.makeKeyAndOrderFront(nil)
-    }
-}
-
-private struct BonsplitTabBarDebugView: View {
-    @AppStorage(BonsplitTabBarDebugSettings.separatorFadeWidthKey)
-    private var separatorFadeWidth = BonsplitTabBarDebugSettings.defaultSeparatorFadeWidth
-    @AppStorage(BonsplitTabBarDebugSettings.contentFadeWidthKey)
-    private var contentFadeWidth = BonsplitTabBarDebugSettings.defaultContentFadeWidth
-    @AppStorage(BonsplitTabBarDebugSettings.solidSurfaceWidthAdjustmentKey)
-    private var solidSurfaceWidthAdjustment = BonsplitTabBarDebugSettings.defaultSolidSurfaceWidthAdjustment
-
-    private var resolvedSeparatorFadeWidth: Double {
-        BonsplitTabBarDebugSettings.resolvedSeparatorFadeWidth(separatorFadeWidth)
-    }
-
-    private var resolvedContentFadeWidth: Double {
-        BonsplitTabBarDebugSettings.resolvedContentFadeWidth(contentFadeWidth)
-    }
-
-    private var resolvedSolidSurfaceWidthAdjustment: Double {
-        BonsplitTabBarDebugSettings.resolvedSolidSurfaceWidthAdjustment(solidSurfaceWidthAdjustment)
-    }
-
-    private var separatorFadeBinding: Binding<Double> {
-        Binding(
-            get: { resolvedSeparatorFadeWidth },
-            set: { setSeparatorFadeWidth($0) }
-        )
-    }
-
-    private var contentFadeBinding: Binding<Double> {
-        Binding(
-            get: { resolvedContentFadeWidth },
-            set: { setContentFadeWidth($0) }
-        )
-    }
-
-    private var solidSurfaceWidthAdjustmentBinding: Binding<Double> {
-        Binding(
-            get: { resolvedSolidSurfaceWidthAdjustment },
-            set: { setSolidSurfaceWidthAdjustment($0) }
-        )
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 14) {
-            Text(String(localized: "debug.bonsplitTabBarDebug.heading", defaultValue: "Bonsplit Tab Bar"))
-                .font(.headline)
-
-            GroupBox(String(localized: "debug.bonsplitTabBarDebug.actionLaneGeometry", defaultValue: "Action Lane Geometry")) {
-                VStack(alignment: .leading, spacing: 10) {
-                    BonsplitTabBarDebugSliderRow(
-                        title: String(localized: "debug.bonsplitTabBarDebug.contentFade", defaultValue: "Content fade"),
-                        value: contentFadeBinding,
-                        setting: BonsplitTabBarDebugSettings.contentFadeWidthSetting
-                    )
-                    BonsplitTabBarDebugSliderRow(
-                        title: String(localized: "debug.bonsplitTabBarDebug.solidBgExtra", defaultValue: "Solid bg extra"),
-                        value: solidSurfaceWidthAdjustmentBinding,
-                        setting: BonsplitTabBarDebugSettings.solidSurfaceWidthAdjustmentSetting
-                    )
-                }
-                .padding(.top, 2)
-            }
-
-            GroupBox(String(localized: "debug.bonsplitTabBarDebug.actionLaneBorder", defaultValue: "Action Lane Border")) {
-                VStack(alignment: .leading, spacing: 10) {
-                    BonsplitTabBarDebugSliderRow(
-                        title: String(
-                            localized: "debug.bonsplitTabBarDebug.separatorFadeFrame",
-                            defaultValue: "Separator fade frame"
-                        ),
-                        value: separatorFadeBinding,
-                        setting: BonsplitTabBarDebugSettings.separatorFadeWidthSetting
-                    )
-                }
-                .padding(.top, 2)
-            }
-
-            HStack(spacing: 10) {
-                Button(String(localized: "debug.bonsplitTabBarDebug.reset", defaultValue: "Reset")) {
-                    setSeparatorFadeWidth(BonsplitTabBarDebugSettings.defaultSeparatorFadeWidth)
-                    setContentFadeWidth(BonsplitTabBarDebugSettings.defaultContentFadeWidth)
-                    setSolidSurfaceWidthAdjustment(BonsplitTabBarDebugSettings.defaultSolidSurfaceWidthAdjustment)
-                }
-                Button(String(localized: "debug.bonsplitTabBarDebug.copyConfig", defaultValue: "Copy Config")) {
-                    BonsplitTabBarDebugSettings.copyCurrentTuningToPasteboard()
-                }
-            }
-
-            Text(verbatim: BonsplitTabBarDebugSettings.currentTuningDescription())
-                .font(.system(.caption, design: .monospaced))
-                .textSelection(.enabled)
-                .lineLimit(3)
-                .frame(maxWidth: .infinity, alignment: .leading)
-        }
-        .padding(16)
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-    }
-
-    private func setSeparatorFadeWidth(_ value: Double) {
-        separatorFadeWidth = BonsplitTabBarDebugSettings.resolvedSeparatorFadeWidth(value)
-        refreshLiveWorkspaces()
-    }
-
-    private func setContentFadeWidth(_ value: Double) {
-        contentFadeWidth = BonsplitTabBarDebugSettings.resolvedContentFadeWidth(value)
-        refreshLiveWorkspaces()
-    }
-
-    private func setSolidSurfaceWidthAdjustment(_ value: Double) {
-        solidSurfaceWidthAdjustment = BonsplitTabBarDebugSettings.resolvedSolidSurfaceWidthAdjustment(value)
-        refreshLiveWorkspaces()
-    }
-
-    private func refreshLiveWorkspaces() {
-        let managers = AppDelegate.shared?.allMainWindowTabManagersForDebug() ?? []
-        var seen = Set<ObjectIdentifier>()
-        for manager in managers {
-            guard seen.insert(ObjectIdentifier(manager)).inserted else { continue }
-            manager.refreshSplitButtonBackdropEffect()
-        }
-    }
-}
-
-private struct BonsplitTabBarDebugSliderRow: View {
-    let title: String
-    @Binding var value: Double
-    let setting: BonsplitTabBarDebugNumberSetting
-
-    private var resolvedValue: Double {
-        setting.resolved(value)
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 6) {
-            HStack(spacing: 8) {
-                Text(title)
-                    .frame(width: 112, alignment: .leading)
-                Slider(
-                    value: Binding(
-                        get: { resolvedValue },
-                        set: { value = setting.resolved($0) }
-                    ),
-                    in: setting.range,
-                    step: setting.step
-                )
-                Text("\(setting.format(resolvedValue)) px")
-                    .font(.caption)
-                    .monospacedDigit()
-                    .frame(width: 76, alignment: .trailing)
-            }
-
-            Stepper(
-                String(localized: "debug.bonsplitTabBarDebug.fineTune", defaultValue: "Fine tune"),
-                value: Binding(
-                    get: { resolvedValue },
-                    set: { value = setting.resolved($0) }
-                ),
-                in: setting.range,
-                step: setting.step
-            )
-        }
-    }
-}
-
 // MARK: - Tab Bar Backdrop Lab Window
 
 private final class TabBarBackdropLabWindowController: NSWindowController, NSWindowDelegate {
@@ -3590,27 +3390,6 @@ private final class TabBarBackdropLabWindowController: NSWindowController, NSWin
         window?.center()
         window?.makeKeyAndOrderFront(nil)
         window?.orderFrontRegardless()
-    }
-}
-
-private struct TabBarBackdropLabVariant: Identifiable {
-    let id: String
-    let title: String
-    let detail: String
-    let effect: BonsplitConfiguration.Appearance.SplitButtonBackdropEffect
-    let chromeHex: String
-    let tabBarHex: String
-    let splitButtonBackdropHex: String
-    let paneHex: String
-    let borderHex: String
-    let terminalColor: NSColor
-    let surfaceColor: NSColor
-    let separatorColor: NSColor
-    let opacity: CGFloat
-
-    var renderIdentity: String {
-        let separatorFadeWidth = effect.separatorFadeWidth.map { String(format: "%.1f", $0) } ?? "nil"
-        return "\(id)-\(chromeHex)-\(tabBarHex)-\(splitButtonBackdropHex)-\(paneHex)-\(borderHex)-\(String(format: "%.3f", opacity))-\(String(format: "%.1f", effect.fadeWidth))-\(String(format: "%.1f", effect.contentFadeWidth))-\(String(format: "%.1f", effect.solidWidth))-\(String(format: "%.1f", effect.solidSurfaceWidthAdjustment))-\(separatorFadeWidth)-\(String(format: "%.2f", effect.fadeRampStartFraction))-\(String(format: "%.2f", effect.leadingOpacity))-\(String(format: "%.2f", effect.trailingOpacity))-\(String(format: "%.2f", effect.contentOcclusionFraction))-\(effect.masksTabContent ? 1 : 0)"
     }
 }
